salesforce_tab_performance/requirements.txt
selenium
pytest
allure-pytest
openpyxl
